requirements.txt
beautifulsoup4==4.12.2
Django==3.1
django-cors-headers==3.11.0
djangorestframework==3.11.1
joblib==0.16.0
numpy==1.18.4
requests==2.29.0
scikit-learn==0.24.2
whois==0.9.27